requirements.txt
asgiref==3.2.10
certifi==2020.6.20
Django==3.0.8
django-cors-headers==3.4.0
gunicorn==20.0.4
pytz==2020.1
sqlparse==0.3.1
whitenoise==5.1.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
PyJWT==1.7.1
